+++ site/homepage/content/release-notes.md
@@ -20,6 +20,7 @@ description = "Information about changes in recent Hono releases. Includes new f
 * The Hono container images released with tag 1.10.0 failed to start up when not running as user `root` because the
   Java process was lacking authority to create a temporary directory in the file system's root folder (`/`).
   This has been fixed.
+* Command response messages published via Kafka did not contain the `tenant_id` header. This has been fixed.
 
 ## 1.10.0